[#11] Replace catch2 with gtest
* Add gtest test_requirement
* Update builder and CMakeLists.txt files

+++ builder/controllers/conan_install.py
@@ -8,14 +8,14 @@ class ConanInstallController(object):
     def __init__(self):
         self.conan = Conan()
 
-    def install_conan_dependencies(self, host_profile, output_folder, clean=False):
+    def install_conan_dependencies(self, host_profile, output_folder, with_tests, clean=False):
         logging.debug(f"Installing Conan dependencies for profile '{host_profile}'...")
         self.__validate_install_dir(Path(output_folder), clean)
         self.conan.check_availability()
-        self.conan.install_dependencies(host_profile, output_folder)
+        self.conan.install_dependencies(host_profile, output_folder, with_tests)
 
     def __validate_install_dir(self, output_folder: Path, clean: bool):
-        if clean:
+        if clean and output_folder.exists():
             shutil.rmtree(output_folder)
             logging.debug(f"Removed '{output_folder}'")
         if clean or not output_folder.exists():
